+++ packages/plugin-react/test/react-i18next.test.ts
@@ -1,5 +1,5 @@
 // eslint-disable-next-line import/no-unresolved
 import jest from '@enum-plus/test/engines/jest';
-import testLocalization from './test-suites/localize';
+import testReactI18next from './test-suites/react-i18next';
 
-testLocalization(jest);
+testReactI18next(jest);